Adds automatic Mnesia table copy type manager.
Fetches the size of vehicle and prediction tables once per second.
(This is useful on somewhat low memory systems when you might stop
mta_fetcher_database temporarily for some reason or other.)

* If the last five table size samples are < 200 items, convert both
  tables to disc_copies.
* Else, if the *two* most recent table size samples are >= 200 items,
  convert both tables to disc_only_copies.
* Else, convert to disc_copies.

If we convert to disc_copies, wait 20 seconds before performing another
conversion.
If we convert to disc_only_copies, wait 5 minutes before performing
another conversion.
If -when we go to convert-, we find that the tables are already of the
target copy type, do nothing and don't wait any additional time before
the next conversion.

200 items puts us between ~300MB and 400MB of resident size on a 64-bit
system. Tweak TABLE_COPY_TRANSITION_SIZE in process.erl to adjust this.
This should *probably* be a configuration parameter.